% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_order.R
\name{partial_order}
\alias{partial_order}
\title{Partial order}
\usage{
partial_order(x, n = length(x), decreasing = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{n}{integer.}

\item{decreasing}{logical.}
}
\value{
integer vector.
}
\description{
function to order partially a numeric vector.
}
\examples{
set.seed(1L)
x <- sample.int(10, 5)
x[partial_order(x, 3)]
}
